requirements.txt
pkg_resources==0.0.0
pyaes==1.6.1
pyasn1==0.4.8
python-dotenv==1.0.0
rsa==4.9
Telethon==1.27.0
